requirements.txt
discord.py
groq
python-dotenv  
chatterbot
spacy